+++ packages/core/textfield/src/types.ts
@@ -30,6 +30,11 @@ export interface PublicProps
   theme?: ThemeProp<ThemeTokens, ThemeProps>;
   /** Mousedown handler that will fire on the container element */
   onMouseDown?: React.MouseEventHandler<HTMLElement>;
+  /**
+   * A `testId` prop is provided for specified elements, which is a unique
+   * string that appears as a data attribute `data-testid` in the rendered code,
+   * serving as a hook for automated tests */
+  testId?: string;
 }
 
 export interface InternalProps extends PublicProps {