+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/nodeviews/mediaNodeUpdater.spec.ts
@@ -24,8 +24,12 @@ describe('MediaNodeUpdater', () => {
     jest
       .spyOn(MediaClientModule, 'getMediaClient')
       .mockReturnValue(mediaClient);
-    jest.spyOn(commands, 'updateMediaNodeAttrs').mockReturnValue(() => true);
-    jest.spyOn(commands, 'replaceExternalMedia').mockReturnValue(() => true);
+    jest.spyOn(commands, 'updateAllMediaNodesAttrs').mockReturnValue(() => {});
+    jest.spyOn(commands, 'updateMediaNodeAttrs').mockReturnValue(() => {});
+    jest.spyOn(commands, 'replaceExternalMedia').mockReturnValue(() => {});
+    jest
+      .spyOn(mediaCommon, 'getViewMediaClientConfigFromMediaProvider')
+      .mockReturnValue(getDefaultMediaClientConfig());
 
     const contextIdentifierProvider: Promise<
       ContextIdentifierProvider
@@ -81,8 +85,8 @@ describe('MediaNodeUpdater', () => {
 
       await mediaNodeUpdater.updateContextId();
 
-      expect(commands.updateMediaNodeAttrs).toBeCalledTimes(1);
-      expect(commands.updateMediaNodeAttrs).toBeCalledWith(
+      expect(commands.updateAllMediaNodesAttrs).toBeCalledTimes(1);
+      expect(commands.updateAllMediaNodesAttrs).toBeCalledWith(
         'source-file-id',
         {
           __contextId: 'object-id',
@@ -118,8 +122,8 @@ describe('MediaNodeUpdater', () => {
           collectionName: 'source-collection',
         },
       );
-      expect(commands.updateMediaNodeAttrs).toBeCalledTimes(1);
-      expect(commands.updateMediaNodeAttrs).toBeCalledWith(
+      expect(commands.updateAllMediaNodesAttrs).toBeCalledTimes(1);
+      expect(commands.updateAllMediaNodesAttrs).toBeCalledWith(
         'source-file-id',
         {
           __fileName: 'some-file',
@@ -163,8 +167,8 @@ describe('MediaNodeUpdater', () => {
           collectionName: 'source-collection',
         },
       );
-      expect(commands.updateMediaNodeAttrs).toBeCalledTimes(1);
-      expect(commands.updateMediaNodeAttrs).toBeCalledWith(
+      expect(commands.updateAllMediaNodesAttrs).toBeCalledTimes(1);
+      expect(commands.updateAllMediaNodesAttrs).toBeCalledWith(
         'source-file-id',
         {
           __fileName: 'some-file',