+++ packages/help/help/src/components/Article/RelatedArticles/index.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { UIAnalyticsEvent } from '@atlaskit/analytics-next';
 
 import { ArticleItem } from '../../../model/Article';
 
@@ -6,23 +7,32 @@ import RelatedArticlesContent from './RelatedArticlesContent';
 import RelatedArticlesLoading from './RelatedArticlesLoading';
 
 interface Props {
+  /* Display loading state. This prop is optional */
   isLoading?: boolean;
+  /* List of related articles. This prop is optional */
   relatedArticles?: ArticleItem[];
+  /* function executed when the user clicks on of the related articles */
+  onRelatedArticlesListItemClick?: (
+    id: string,
+    analyticsEvent: UIAnalyticsEvent,
+  ) => void;
 }
 
 const RelatedArticles: React.SFC<Props> = (props: Props) => {
-  const { relatedArticles, isLoading } = props;
+  const {
+    isLoading = false,
+    relatedArticles = [],
+    onRelatedArticlesListItemClick,
+  } = props;
 
   return isLoading ? (
     <RelatedArticlesLoading />
   ) : (
-    <RelatedArticlesContent relatedArticles={relatedArticles} />
+    <RelatedArticlesContent
+      relatedArticles={relatedArticles}
+      onRelatedArticlesListItemClick={onRelatedArticlesListItemClick}
+    />
   );
-};
-
-RelatedArticles.defaultProps = {
-  isLoading: false,
-  relatedArticles: [],
 };
 
 export default RelatedArticles;