+++ packages/navigation/atlassian-switcher/docs/04-available-products-custom-provider.tsx
@@ -1,4 +1,5 @@
-import { md, code } from '@atlaskit/docs';
+import * as React from 'react';
+import { md, code, Example } from '@atlaskit/docs';
 
 export default md`
 
@@ -40,4 +41,18 @@ const customAvailableProductsDataProvider = createAvailableProductsProvider(
     </Button>
   </AtlassianSwitcherPrefetchTrigger>
   `}
+
+  ### Example
+
+  ${(
+    <Example
+      packageName="@atlaskit/atlassian-switcher"
+      Component={
+        require('../examples/42-with-prefetch-and-custom-provider').default
+      }
+      title="Switcher with prefetch and custom provider"
+      source={require('!!raw-loader!../examples/42-with-prefetch-and-custom-provider')}
+    />
+  )}
+
   `;