+++ packages/core/calendar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/calendar
 
+## 9.0.0
+
+### Major Changes
+
+- [major][8c725d46ec](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8c725d46ec):
+
+  @atlaskit/calendar has been converted to Typescript. Typescript consumers will now get static type safety. Flow types are no longer provided. No API or behavioural changes.
+
 ## 8.1.0
 
 ### Minor Changes