+++ packages/people-and-teams/profilecard/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/profilecard
 
+## 12.1.0
+
+### Minor Changes
+
+- [minor][16bf116576](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/16bf116576):
+
+  Increased show/hide delay times for ProfileCardTrigger
+
 ## 12.0.9
 
 ### Patch Changes